refactor: finding and updating user logic and validate req param

+++ src/modules/users/users.controller.js
@@ -1,34 +1,25 @@
 import * as userService from './users.service.js';
-import User from './users.model.js';
-import { getOrSet } from '../../utils/cache.js';
 
 export async function getUser(req, res, _next) {
     const { userId } = req.params;
     if (!userId) {
-        return res.status(400).json({ message: 'User ID is required' });
+        return res.status(400).json({ message: 'User ID param is required' });
     }
 
-    try {
-        const userWithoutPassword = await getOrSet(
-            `user:${userId}`,
-            async () => {
-                const userRow = await User.findByPk(userId, { attributes: { exclude: ['password'] } });
-                return userRow?.dataValues;
-            }
-        );
-        if (!userWithoutPassword) {
-            return res.status(404).json({ message: 'User not found' });
-        }
-
-        res.status(200).json({ user: userWithoutPassword });
-    } catch {
-        throw new Error;
+    const userWithoutPassword = await userService.findUser(userId);
+    if (!userWithoutPassword) {
+        return res.status(404).json({ message: 'User not found' });
     }
+
+    res.status(200).json({ user: userWithoutPassword });
 }
 
 export async function updateUser(req, res, _next) {
     const { userId } = req.params;
     const { body: userData } = req;
+    if (!userId) {
+        return res.status(400).json({ message: 'User ID param is required' });
+    }
 
     const userWithoutPassword = await userService.updateUser(userId, userData);
     if (!userWithoutPassword) {
@@ -41,6 +32,9 @@ export async function updateUser(req, res, _next) {
 export async function deleteUser(req, res, _next) {
     const { userId: deletedUserId } = req.params;
     const { body: { userId, password } } = req;
+    if (!deletedUserId) {
+        return res.status(400).json({ message: 'User ID param is required' });
+    }
 
     const isDeleted = await userService.deleteUser(userId, password, deletedUserId);
     if (!isDeleted) {
